[Refactor] 메인페이지 디렉토리 추가 및 관련 페이지 이동
