+++ front/src/App.js
@@ -8,13 +8,13 @@ import Loading from './component/common/Loading';
 import Header from './component/common/header/Header';
 import LoginModal from './component/common/modal/LoginModal';
 import 'react-toastify/dist/ReactToastify.css';
-import RestaurantPage from './pages/RestaurantPage';
-import StayPage from './pages/StayPage';
-import SpotPage from './pages/SpotPage';
+import RestaurantPage from './pages/mainPages/RestaurantPage';
+import StayPage from './pages/mainPages/StayPage';
+import SpotPage from './pages/mainPages/SpotPage';
 import Footer from './component/common/footer/Footer';
-import AllThemePage from './pages/AllThemePage';
+import AllThemePage from './pages/mainPages/AllThemePage';
 
-const MainPage = React.lazy(() => import('./pages/MainPage'));
+const MainPage = React.lazy(() => import('./pages/mainPages/MainPage'));
 const PublishPage = React.lazy(() => import('./pages/PublishPage'));
 const PostDetailPage = React.lazy(() => import('./pages/PostDetailPage'));
